feat: Extract AppTile
