+++ react/AppTile/helpers.js
