+++ react/AppTile/styles.styl
@@ -22,6 +22,7 @@
   align-items center
   margin-top .5rem
   text-align center
+  max-width 100%
 
 .AppTile-title, .AppTile-developer, .AppTile-status
   display block
